refactor: rename provider to context in tasks components
Replaced `provider` with `context` across tasks and inbox components to ensure consistency with established naming conventions. Updated all related imports, exports, file paths, and references to reflect this change.
