+++ apps/nextjs/src/components/features/inbox/inbox-list-item/workspace-activity/task-list-item/action-button/action-button.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { useInboxContext } from '@/components/features/inbox/inbox/provider';
+import { useInboxContext } from '@/components/features/inbox/inbox/context';
 import type { IconButtonProps } from '@/components/ui/icon-button';
 import { ArchiveButton } from './archive-button';
 import { MoveToInboxButton } from './move-to-inbox-button';
