+++ apps/nextjs/src/components/features/tasks/tasks-list/index.ts
@@ -1,4 +1,4 @@
-export { useTasksListContext } from './provider';
+export { useTasksListContext } from './context';
 export * from './tasks-list';
 export * from './tasks-list-body';
 export * from './tasks-list-content';
