+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-name/index.ts
@@ -1,6 +1,6 @@
 export * from './tasks-name';
 export * from './tasks-name-cell';
+export * from './tasks-name-context';
 export * from './tasks-name-field';
 export * from './tasks-name-grab-icon';
-export * from './tasks-name-provider';
 export * from './tasks-name-right-container';
