+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-name/task-parent-name/task-parent-name.tsx
@@ -4,7 +4,7 @@ import { Flex } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
 import { Text } from '@/components/ui/text';
 import { useTask } from '@/store/entities/task';
-import { useTasksNameContext } from '../tasks-name-provider';
+import { useTasksNameContext } from '../tasks-name-context';
 
 export const TaskParentName = memo(function TaskParentName() {
   const { taskId } = useTasksNameContext();
