+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-name/tasks-name-grab-icon-container.tsx
@@ -1,7 +1,7 @@
 import type React from 'react';
 import { memo, useMemo } from 'react';
 import { Flex, type FlexProps } from '@/components/ui/flex';
-import { useTasksNameContext } from './tasks-name-provider';
+import { useTasksNameContext } from './tasks-name-context';
 
 type Props = FlexProps;
