+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-name/tasks-name-right-container.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 import { Flex, type FlexProps } from '@/components/ui/flex';
-import { useTasksNameContext } from './tasks-name-provider';
+import { useTasksNameContext } from './tasks-name-context';
 
 type Props = FlexProps;
