+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-header/tasks-list-header.tsx
@@ -3,7 +3,7 @@ import { useTasksTaskColumnIds } from '@/components/features/tasks/hooks';
 import { Flex } from '@/components/ui/flex';
 import { Column } from './column';
 import { RemainingSpace } from './columns';
-import { Context, useTasksListHeaderContext } from './provider';
+import { Context, useTasksListHeaderContext } from './context';
 
 export const TasksListHeader = memo(function TasksListHeader() {
   return (
