+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-section/index.ts
@@ -1,5 +1,5 @@
 export {
   TasksListSectionContext,
   useTasksListSectionContext,
-} from './provider';
+} from './context';
 export * from './tasks-list-section';
